OvmfPkg: Replace legacy 8254 PIT with local APIC timer
Instead of relying on the legacy 8254 PIT as the timer UEFI services,
rely on the XenTimerDxe implementation. This is not Xen specific, as it
simply uses the local APIC timer triggering interrupts on the vector 32.

This is a more generic approach that works for both QEMU and Cloud
Hypervisor as it doesn't require the VMM to emulate a 8254 PIT.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>